don't use char[], and use wordexp on DEFAULT_DIR
wordexp allows us to have the default of ~/Music, a sane default that
should work for most linux users.

char arrays (compared to char*s) are mutable, and stored in a
different part of the executable.  I don't know how I got in the
questionable habit of using them, but as I find them I'll switch them
out.